requirements.txt
tabula-py
matplotlib